+++ src/main/java/com/hikers/hikemate/repository/CourseRepository.java
@@ -5,7 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CourseRepository extends JpaRepository<Course, Long> {
-
-
+public interface CourseRepository extends JpaRepository<Course, Integer> {
 }
